Initial commit
Created the ability to draw a path. Draw's a circle which is the sprite's "Range" that shrinks as it gets closer to the moving edge.

TODO:
Update the current line length to calculate based off of all points instead of newest to first.
Make it so that the circle won't grow when click is held further from range.