Path building
Created path building which when the object is selected (default char 1 with esc), it will show a circle which is the distance it can still travel. Then it can be dragged along a saved path. TODO: Make animation for the character to follow that path. Make it so that other chars are selectable. Finish menu buttons.